backend/requirements.txt
# VisioNova Backend Dependencies
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=4.9.0
duckduckgo-search>=6.0.0
